_01/src/test/java/inflearn/yhk/_01/repository/MemoryMemberRepositoryTest.java
package inflearn.yhk._01.repository;

import inflearn.yhk._01.domain.Member;
import static org.assertj.core.api.Assertions.*;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.Test;

import java.util.List;


class MemoryMemberRepositoryTest {

    MemoryMemberRepository repository = new MemoryMemberRepository();


    /**
     * clear object on each test finishes
     */
    @AfterEach
    public void afterEach() {
        repository.clearStore();
    }

    @Test
    public void save() {
        Member member = new Member();
        member.setName("spring");
        repository.save(member);

        Member result = repository.findById(member.getId()).get();

        // assertj (org.assertj.core.api) 😊 👍
        assertThat(member).isEqualTo(result);

        // junit (org.junit.jupiter.api) approach: 😊 👍
        /* Assertions.assertEquals(member, result); */

        // sysout: 😵 👎
        /* System.out.println("result: " + (result == member)); */
    }

    @Test
    public void findByName() {
        Member member1 = new Member();
        member1.setName("spring1");
        repository.save(member1);

        Member member2 = new Member();
        member2.setName("spring2");
        repository.save(member2);

        Member result = repository.findByName("spring1").get();
        assertThat(result).isEqualTo(member1);
    }

    @Test
    public void findAll() {
        Member member1 = new Member();
        member1.setName("spring1");
        repository.save(member1);

        Member member2 = new Member();
        member2.setName("spring2");
        repository.save(member2);

        List<Member> result = repository.findAll();
        assertThat(result.size()).isEqualTo(2);
    }

}
